third_party/WebKit/LayoutTests/typedcssom/computedstyle/border-top-width.html

<!DOCTYPE html>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>

<div id="testElement"></div>

<script>

var computedStyleMap = testElement.computedStyleMap();

test(function() {
  testElement.style.borderTop = '10px solid #000';

  var result = computedStyleMap.get('border-top-width');
  assert_equals(result.constructor.name, CSSUnitValue.name);
  assert_equals(result.toString(), '10px');
}, "Getting a 10px border-top-width returns a CSSUnitValue");

test(function() {
  testElement.style.borderTop = '20px solid #000';

  var result = computedStyleMap.getAll('border-top-width');
  assert_equals(result.length, 1);
  assert_equals(result[0].toString(), '20px');
}, "getAll for border-top-width returns a single value");

</script>
